+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Release history for s390-tools (MIT version)
   For Linux kernel version: 5.x
 
   Changes of existing tools:
+  - dbginfo: Gather qdisc related data (using 'tc')
   - dbginfo: Gather extended network statistics (using 'ip link')
 
   Bug Fixes: